+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Compile SDK version updated to 30
 - Minimum SDK version updated to 21
 - Manifest components updated for Android 12.
+- Removed deprecated method `initialise(Application, Logger.LogLevel, Long)`
 
 ## 3.0.0 (08-05-2021)
 - Support for Lifecycle Callback `2.3.1` and above